Servicos-Projeto-Final/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-ai[groq,openai]
openai
httpx
python-dotenv
